BIG BOSS LEVEL COMPLETE
This is the bug in the code (master) that is causing in consitent results due to reading beyond array and thus garbage.
This is huge! In some cases this might be non heap data and thus give some consistency - which is why it works for subarrays.
For fullframe the extended array size must change stack organization.
This obviously gives diff results when threaded to when not.
Using --fstack-protector-all may have saved this from time to time, either by the extended read reading the canary itself or...

This makes it almost impossible to validate new code against old without changing master.

This bug means that it is possible for results to change from compilation to compilation and/or run to run, when the following is is changed
code, optimization, compiler, os, chip-set, processor optimizer, threaded.

Signed-off-by: James Noss <jnoss@stsci.edu>